use aral_by_size when we actually agreegate stats to aral by size

+++ src/libnetdata/dictionary/dictionary.h
@@ -98,9 +98,9 @@ struct dictionary_stats {
 
     // memory
     struct {
-        long index;                 // bytes of keys indexed (indication of the index size)
-        long values;                // bytes of caller structures
-        long dict;                  // bytes of the structures dictionary needs
+        ssize_t index;              // bytes of keys indexed (indication of the index size)
+        ssize_t values;             // bytes of caller structures
+        ssize_t dict;               // bytes of the structures dictionary needs
     } memory;
 
     // spin locks
